+++ Libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ryusukeyashiro <ryusukeyashiro@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/26 16:55:08 by ryusukeyash       #+#    #+#             */
-/*   Updated: 2024/04/27 17:27:12 by ryyashir         ###   ########.fr       */
+/*   Updated: 2024/12/12 10:36:36 by ryusukeyash      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ int					ft_memcmp(const void *s1, const void *s2, size_t n);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 int					ft_tolower(int c);
 int					ft_toupper(int c);
-int					ft_atoi(const char *str);
+long				ft_atoi(const char *str);
 
 size_t				ft_strlen(const char *str);
 size_t				ft_strlcat(char *dest, const char *src, size_t size);
